app/src/main/java/com/example/storyapp/ui/fragment/login/LoginViewModel.kt
package com.example.storyapp.ui.fragment.login

import android.util.Log
import androidx.lifecycle.LiveData
import androidx.lifecycle.MutableLiveData
import androidx.lifecycle.ViewModel
import com.example.storyapp.api.ApiConfig
import com.example.storyapp.response.MessageLogin
import com.example.storyapp.response.ResponseLogin
import com.google.gson.Gson
import retrofit2.Call
import retrofit2.Callback
import retrofit2.Response

class LoginViewModel : ViewModel() {

    private val _loginResult = MutableLiveData<ResponseLogin>()
    val loginResult: LiveData<ResponseLogin> = _loginResult

    private val _message = MutableLiveData<String>()
    val message: LiveData<String> = _message

    private val _isLoading = MutableLiveData<Boolean>()
    val isLoading : LiveData<Boolean> = _isLoading

    fun loginUser(email: String, password: String){
        _isLoading.value = true
        val client = ApiConfig.getApiService().login(email, password)
        client.enqueue(object : Callback<ResponseLogin>{
            override fun onResponse(
                call: Call<ResponseLogin>,
                response: Response<ResponseLogin>
            ) {
                _isLoading.value = false
                val responseBody = response.body()
                if (response.isSuccessful && responseBody != null){
                    _loginResult.value = responseBody!!
                    _message.value = responseBody.message
                }
                else{
                    val responseError = Gson().fromJson(response.errorBody()?.charStream(),MessageLogin::class.java)
                    _message.value = responseError.message
                    Log.e(TAG, "onResponse2: ${response.message()}")
                }
            }

            override fun onFailure(call: Call<ResponseLogin>, t: Throwable) {
                _isLoading.value = false
                Log.e(TAG, "onResponse3: ${t.message}")
            }

        })
    }


    companion object{
        private const val TAG = "LoginViewModel"
    }
}